Move down build file for web.
Moved batyr-web.opam moved to the root for opam pinning to work.  For
topkg lint to see it, topkg now needs to be run from the root (".." is
not allowed in the path).  Then _tags needs to be moved to the root,
since topkg expects it there, which means ocamlbuild needs to be run
from the root, which means myocamlbuild.ml needs to be moved to the
root.